Feat/add follower followcnt redis (#18)
* 合并冲突

* 合并relation/service/query_friend_list冲突，保留队友的写法

* 将follower_cnt和follow_cnt字段从数据库user表分离增加了redis后，同步修改了query_user_info和register的bug

* 修改了api/relation用c.Query获取user_id的方式, 修改为c.GetInt64获取user_id

* 修改了entity/QueryIsFollow, 如果找不到返回false, nil

* 修复了entity/CreateFollowing 在取关之后进行软删除，再次关注无法创建的bug

* 将user_id的限制从大于0改为大于等于0

* 将查询userList的方法合并到user_service中供其他service调用

* 将relation_service/query_friend_list获取followercnt及followcnt的方法更新

* 安装golangci-lint

* add some error check

* fix some golangci-lint error

+++ internal/service/message_service/message.go
@@ -26,7 +26,7 @@ var messageDao = entity.GetMessageDaoInstance()
 // @return	err			error	当执行出现错误时返回error，否则返回nil
 func SendMessage(fromUserId uint, toUserId uint, content string) error {
 	// 验证toUserId的合法性
-	_, err := userDao.QueryUserById(toUserId)
+	_, err := userDao.QueryUserById(int64(toUserId))
 	if err != nil {
 		return errors.New("消息接收方id不存在")
 	}
